requirements-test.txt
mock
requests-mock
pytest
pytest-pylint
pytest-cov
pubtools-pyxis
pyrsistent==0.15.7;python_version<'3.0'
pushcollector
